Update tkinter_box
No shame, I am a noob, this new commit is an improvement on the last one (9af088f), all thanks to D Hank for input, learning so much! Here’s the major changes:

-no \ needed here, quick and easy to clean it up
"line 33, you can hit enter for the text to wrap to the next line since the length of the line is too long. If it gives you issues, a \ will escape the newline and python will view it all as one line still"
-removed some spaces around line 123
-removed some blank lines
-fixed datetime var in ‘def open_about():” on line 72

Now over to Sentdex for some framing!