ledger update
Many updates here since the last commit! I have managed to build a timestamp that actually takes the current time down to the second at the time when each function is performed. Then I was able to combine all session logging items into one line (function ID and description of action).  Lines 200 - 215 contains session log checking and file creation if condition isn't met. def gen_log() on line 200 generates the CSV with the header row and keeps the file ready for when logging events are to be appended to the file. I have tried appending the exact data that is printed the def view_command in lines 59-62 but failed, any advice is greatly appreciated!